+++ public/pages/Indices/containers/IndicesActions/IndicesActions.test.tsx
@@ -274,7 +274,7 @@ describe("<IndicesActions /> spec", () => {
     await waitFor(() => {
       expect(browserServicesMock.commonService.apiCaller).toHaveBeenCalledTimes(2);
       expect(coreServicesMock.notifications.toasts.addSuccess).toHaveBeenCalledTimes(1);
-      expect(coreServicesMock.notifications.toasts.addSuccess).toHaveBeenCalledWith("Delete successfully");
+      expect(coreServicesMock.notifications.toasts.addSuccess).toHaveBeenCalledWith("Delete [test_index] successfully");
       expect(onDelete).toHaveBeenCalledTimes(1);
     });
   });
